Create EDA (box plots)
I create the boxplots for each predictor of salary. And attempted to make a model with all plots at the bottom.
Almost all have high outliers/ right skewed. 

I know Clayton mentioned a transformation of log110 or sqrt. Is it possible to also subset the data where it's more "normal"?